Changing transfer allowance to decimal in dto

+++ src/SFA.DAS.LevyTransferMatching.Infrastructure/Dto/AccountDto.cs
@@ -2,6 +2,6 @@
 {
     public class AccountDto
     {
-        public double RemainingTransferAllowance { get; set; }
+        public decimal RemainingTransferAllowance { get; set; }
     }
 }
